added Game interface, GameImpl class, and modified affected files
