+++ mgl-webapp/src/main/java/com/organization/mvcproject/model/Company.java
@@ -1,7 +1,6 @@
 
 package com.organization.mvcproject.model;
 
-import java.util.ArrayList;
 import java.util.List;
