+++ mgl-webapp/src/main/java/com/organization/mvcproject/service/GameService.java
@@ -10,5 +10,9 @@ public interface GameService {
 	List<Game> retrieveAllGames();
 
 	Game saveGame(Game game);
+	
+	boolean updateGame(Game game);
+	
+	boolean deleteGame(Game game);
 
 }
